+++ 1_Object_Oriented/.editorconfig
@@ -322,3 +322,13 @@ trim_trailing_whitespace = false
 
 [*/Migrations/*]
 generated_code = true
+
+# Verify settings
+[*.{received,verified}.{json,txt,xml}]
+charset = "utf-8-bom"
+end_of_line = lf
+indent_size = unset
+indent_style = unset
+insert_final_newline = false
+tab_width = unset
+trim_trailing_whitespace = false
