+++ 1_Object_Oriented/test/IntegrationTests/Alexandria.Local.IntegrationTests/GlobalSuppressions.cs
@@ -15,3 +15,10 @@ using System.Diagnostics.CodeAnalysis;
     "CA1515:Consider making public types internal",
     Justification = "Allowed for test project"
 )]
+[assembly: SuppressMessage(
+    "Naming",
+    "CA1707:Identifiers should not contain underscores",
+    Justification = "Allowed for test project",
+    Scope = "namespaceanddescendants",
+    Target = "~N:Alexandria.Local.IntegrationTests.Tests"
+)]
