+++ 1_Object_Oriented/test/IntegrationTests/Alexandria.Local.IntegrationTests/GlobalUsing.cs
@@ -3,6 +3,5 @@
 global using static Alexandria.Local.IntegrationTests.Support.KiotaExtensions;
 global using Aspire.Hosting.ApplicationModel;
 global using Aspire.Hosting.Testing;
-global using FluentAssertions;
 //global using System.Net;
 global using Microsoft.Extensions.DependencyInjection;
